# api/swagger/swagger.yaml
definitions:
  controllers.StartAgentRequest:
    properties:
      role:
        type: string
      task:
        type: string
    required:
    - role
    - task
    type: object
...
host: localhost:8080
info:
  contact:
    email: roackb2@gmail.com
    name: Jay / Fienna Liang
    url: https://github.com/roackb2
  description: This is the API for the Lucid project.
  title: Lucid API
  version: "1.0"
paths:
  /api/v1/agents/create:
    post:
      consumes:
      - application/json
      description: Starts a new agent with role and task
      parameters:
      - description: Agent details
        in: body
        name: agent
        required: true
        schema:
          $ref: '#/definitions/controllers.StartAgentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Agent created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a new agent
      tags:
      - agents
...
securityDefinitions:
  None:
    type: basic
swagger: "2.0"